third_party/WebKit/LayoutTests/editing/pasteboard/paste-text-with-style-6.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
  assert_selection(
    [
      '<div style="font-size: 50px;">',
        '<style>.inner { font-size: 50%; }</style>',
        '<div class="inner" contenteditable>^foo|</div>',
      '</div>'
    ].join(''),
    selection => {
      const doc = selection.document;
      doc.execCommand('Copy');
      selection.modify('move', 'forward', 'character');
      doc.execCommand('Paste');
    },
    [
      '<div style="font-size: 50px;">',
        '<style>.inner { font-size: 50%; }</style>',
        '<div class="inner" contenteditable>foofoo|</div>',
      '</div>'
    ].join(''));
}, 'Copy and paste should not append style');
</script>
